# [sonic_py_common] Cache Static Information in device_info to speed up CLI response