--- a/src/sonic-py-common/sonic_py_common/device_info.py
+++ b/src/sonic-py-common/sonic_py_common/device_info.py
@@ -39,6 +39,10 @@ CHASSIS_INFO_SERIAL_FIELD = 'serial'
 CHASSIS_INFO_MODEL_FIELD = 'model'
 CHASSIS_INFO_REV_FIELD = 'revision'
 
+# Cacheable Objects
+sonic_ver_info = None
+is_chassis_type = None
+
 def get_localhost_info(field, config_db=None):
     try:
         # TODO: enforce caller to provide config_db explicitly and remove its default value
@@ -333,14 +337,17 @@ def get_sonic_version_info():
     if not os.path.isfile(SONIC_VERSION_YAML_PATH):
         return None
 
-    data = {}
+    global sonic_ver_info
+    if sonic_ver_info:
+        return sonic_ver_info
+
     with open(SONIC_VERSION_YAML_PATH) as stream:
         if yaml.__version__ >= "5.1":
-            data = yaml.full_load(stream)
+            sonic_ver_info = yaml.full_load(stream)
         else:
-            data = yaml.load(stream)
+            sonic_ver_info = yaml.load(stream)
 
-    return data
+    return sonic_ver_info
 
 def get_sonic_version_file():
     if not os.path.isfile(SONIC_VERSION_YAML_PATH):
@@ -437,7 +444,10 @@ def is_packet_chassis():
 
 
 def is_chassis():
-    return is_voq_chassis() or is_packet_chassis()
+    global is_chassis_type
+    if is_chassis_type is None:
+        is_chassis_type = is_voq_chassis() or is_packet_chassis()
+    return is_chassis_type
 
 
 def is_supervisor():