--- a/src/sonic-py-common/tests/device_info_test.py
+++ b/src/sonic-py-common/tests/device_info_test.py
@@ -52,6 +52,32 @@ EXPECTED_GET_MACHINE_INFO_RESULT = {
     'onie_kernel_version': '4.10.11'
 }
 
+SONIC_VERISON_YML = """\
+---
+build_version: 'test_branch.1-a8fbac59d'
+debian_version: '11.4'
+kernel_version: '5.10.0-12-2-amd64'
+asic_type: mellanox
+asic_subtype: 'mellanox'
+commit_id: 'a8fbac59d'
+branch: 'test_branch'
+release: 'master'
+libswsscommon: 1.0.0
+sonic_utilities: 1.2"""
+
+SONIC_VERISON_YML_RESULT = {
+    'build_version': 'test_branch.1-a8fbac59d', 
+    'debian_version': '11.4', 
+    'kernel_version': '5.10.0-12-2-amd64', 
+    'asic_type': 'mellanox', 
+    'asic_subtype': 'mellanox', 
+    'commit_id': 'a8fbac59d', 
+    'branch': 'test_branch', 
+    'release': 'master', 
+    'libswsscommon': '1.0.0', 
+    'sonic_utilities': 1.2
+}
+
 class TestDeviceInfo(object):
     @pytest.fixture(scope="class", autouse=True)
     def sanitize_environment(self):
@@ -82,6 +108,40 @@ class TestDeviceInfo(object):
                      "model": SonicV2Connector.TEST_MODEL, 
                      "revision": SonicV2Connector.TEST_REV}
             assert result == truth
+
+    def test_get_sonic_version(self):
+        with mock.patch("os.path.isfile") as mock_isfile:
+            mock_isfile.return_value = True
+            open_mocked = mock.mock_open(read_data=SONIC_VERISON_YML)
+            with mock.patch("{}.open".format(BUILTINS), open_mocked):
+                for _ in range(0,5):
+                    assert device_info.get_sonic_version_info() == SONIC_VERISON_YML_RESULT
+                # Assert the file was read only once
+                open_mocked.assert_called_once_with(device_info.SONIC_VERSION_YAML_PATH)
+        
+    @mock.patch("sonic_py_common.device_info.get_platform_info")
+    def test_is_chassis(self, mock_platform_info):
+        mock_platform_info.return_value = {"switch_type": "npu"}
+        for _ in range(0,10):
+            assert device_info.is_chassis() == False
+        # assert that that get_platform_info is only hit twice
+        assert mock_platform_info.call_count == 2
+        assert device_info.is_voq_chassis() == False
+        assert device_info.is_packet_chassis() == False
+
+        # Clear Cache
+        device_info.is_chassis_type = None
+        mock_platform_info.return_value = {"switch_type": "voq"}
+        assert device_info.is_voq_chassis() == True
+        assert device_info.is_packet_chassis() == False
+        assert device_info.is_chassis() == True
+
+        # Clear Cache
+        device_info.is_chassis_type = None
+        mock_platform_info.return_value = {"switch_type": "chassis-packet"}
+        assert device_info.is_voq_chassis() == False
+        assert device_info.is_packet_chassis() == True
+        assert device_info.is_chassis() == True
 
     @classmethod
     def teardown_class(cls):