--- a/src/sonic-py-common/tests/device_info_test.py
+++ b/src/sonic-py-common/tests/device_info_test.py
@@ -122,24 +122,45 @@ class TestDeviceInfo(object):
     @mock.patch("sonic_py_common.device_info.get_platform_info")
     def test_is_chassis(self, mock_platform_info):
         mock_platform_info.return_value = {"switch_type": "npu"}
-        for _ in range(0,10):
-            assert device_info.is_chassis() == False
-        # assert that that get_platform_info is only hit twice
-        assert mock_platform_info.call_count == 2
+        assert device_info.is_chassis() == False
         assert device_info.is_voq_chassis() == False
         assert device_info.is_packet_chassis() == False
 
-        device_info.is_chassis_type = None
         mock_platform_info.return_value = {"switch_type": "voq"}
         assert device_info.is_voq_chassis() == True
         assert device_info.is_packet_chassis() == False
         assert device_info.is_chassis() == True
 
-        device_info.is_chassis_type = None
         mock_platform_info.return_value = {"switch_type": "chassis-packet"}
         assert device_info.is_voq_chassis() == False
         assert device_info.is_packet_chassis() == True
         assert device_info.is_chassis() == True
+
+        mock_platform_info.return_value = {}
+        assert device_info.is_voq_chassis() == False
+        assert device_info.is_packet_chassis() == False
+        assert device_info.is_chassis() == False
+
+    @mock.patch("sonic_py_common.device_info.ConfigDBConnector", autospec=True)
+    @mock.patch("sonic_py_common.device_info.get_sonic_version_info")
+    @mock.patch("sonic_py_common.device_info.get_machine_info")
+    @mock.patch("sonic_py_common.device_info.get_hwsku")
+    def test_get_platform_info(self, mock_hwsku, mock_machine_info, mock_sonic_ver, mock_cfg_db):
+        mock_cfg_inst = mock_cfg_db.return_value
+        mock_cfg_inst.get_table.return_value = {"localhost": {"switch_type": "npu"}}
+        mock_sonic_ver.return_value = SONIC_VERISON_YML_RESULT
+        mock_machine_info.return_value = {"onie_platform" : "x86_64-mlnx_msn2700-r0"}
+        mock_hwsku.return_value = "Mellanox-SN2700"
+        for _ in range(0,5):
+            hw_info_dict = device_info.get_platform_info()
+            assert hw_info_dict["asic_type"] == "mellanox"
+            assert hw_info_dict["platform"] == "x86_64-mlnx_msn2700-r0"
+            assert hw_info_dict["hwsku"] == "Mellanox-SN2700"
+            assert hw_info_dict["switch_type"] == "npu"
+        assert mock_sonic_ver.called_once()
+        assert mock_machine_info.called_once()
+        assert mock_hwsku.called_once()
+        mock_cfg_inst.get_table.assert_called_once_with("DEVICE_METADATA")
 
     @classmethod
     def teardown_class(cls):