@yxieca 2022-08-27
@vivekrnv this PR causes build failure in 202205 branch. Please raise separate if needed for 202205 branch.